refactor: weekly report 관심사 분리
