+++ src/main/java/bsise/server/report/weekly/service/WeeklyReportService.java
@@ -1,27 +1,25 @@
-package bsise.server.report;
+package bsise.server.report.weekly.service;
 
-import bsise.server.clovar.ClovaResponseDto;
-import bsise.server.clovar.ClovaService;
-import bsise.server.clovar.dailyReport.ClovaDailyAnalysisResult;
-import bsise.server.clovar.dailyReport.DailyReportExtractor;
-import bsise.server.error.DailyReportNotFoundException;
-import bsise.server.error.DuplicateDailyReportException;
+import bsise.server.clova.dto.ClovaResponseDto;
+import bsise.server.clova.service.ClovaService;
+import bsise.server.clova.dailyReport.ClovaDailyAnalysisResult;
+import bsise.server.clova.dailyReport.DailyReportExtractor;
 import bsise.server.error.DuplicationWeeklyReportException;
-import bsise.server.error.LetterNotFoundException;
 import bsise.server.error.WeeklyReportNotFoundException;
 import bsise.server.letter.Letter;
 import bsise.server.letter.LetterRepository;
-import bsise.server.report.daily.DailyReportDto;
-import bsise.server.report.daily.DailyReportRepository;
-import bsise.server.report.daily.DailyReportResponseDto;
-import bsise.server.report.daily.LetterAnalysisRepository;
-import bsise.server.report.weekly.WeeklyReportRepository;
+import bsise.server.report.daily.domain.CoreEmotion;
+import bsise.server.report.daily.domain.DailyReport;
+import bsise.server.report.daily.domain.LetterAnalysis;
+import bsise.server.report.daily.repository.DailyReportRepository;
+import bsise.server.report.daily.repository.LetterAnalysisRepository;
+import bsise.server.report.weekly.domain.WeeklyReport;
 import bsise.server.report.weekly.dto.ClovaWeeklyReportRequestDto;
 import bsise.server.report.weekly.dto.WeeklyReportRequestDto;
 import bsise.server.report.weekly.dto.WeeklyReportResponseDto;
+import bsise.server.report.weekly.repository.WeeklyReportRepository;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
-import java.time.LocalTime;
 import java.util.List;
 import java.util.Map;
 import java.util.Map.Entry;
@@ -32,11 +30,10 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-
 @Service
 @Transactional
 @RequiredArgsConstructor
-public class ReportService {
+public class WeeklyReportService {
 
     private final DailyReportRepository dailyReportRepository;
     private final LetterRepository letterRepository;
@@ -45,111 +42,8 @@ public class ReportService {
     private final WeeklyReportRepository weeklyReportRepository;
 
     /**
-     * <ol> 이 메서드는 순차대로 아래 작업을 수행합니다.
-     *     <li>전달받은 유저 아이디와 대상 날짜에 해당하는 일일 리포트가 이미 존재하는지 확인합니다.</li>
-     *     <li>가장 최근 편지 3개를 찾아 분석합니다. 오늘이라면 현재 시점 기준, 오늘 이전이라면 해당 날짜 기준 가장 최근 편지 3개를 조회합니다.</li>
-     *     <li>클로바에게 조회된 편지로 일일 리포트를 생성을 요청합니다.</li>
-     *     <li>분석된 일일 리포트를 저장하고 응답합니다.</li>
-     * </ol>
-     *
-     * @param dailyReportDto 일일 리포트 생성 요청 DTO
-     * @return 생성된 일일 리포트에 대한 응답 DTO
-     */
-    public DailyReportResponseDto createDailyReport(DailyReportDto.CreateRequest dailyReportDto) {
-        UUID userId = UUID.fromString(dailyReportDto.getUserId());
-        LocalDate targetDate = dailyReportDto.getDate();
-
-        if (dailyReportRepository.existsByUserAndTargetDate(userId, targetDate)) {
-            throw new DuplicateDailyReportException("Duplicate daily report exists.");
-        }
-
-        List<Letter> letters = findRecentLetters(userId, targetDate);
-
-        ClovaResponseDto clovaResponse = requestClovaAnalysis(letters);
-
-        ClovaDailyAnalysisResult clovaDailyAnalysisResult = DailyReportExtractor.extract(clovaResponse);
-        DailyReport dailyReport = buildDailyReport(targetDate, clovaDailyAnalysisResult);
-        dailyReportRepository.save(dailyReport);
-
-        List<LetterAnalysis> letterAnalyses = buildLetterAnalyses(letters, clovaDailyAnalysisResult);
-        letterAnalyses.forEach(analysis -> analysis.getLetter().setDailyReport(dailyReport));
-        letterAnalysisRepository.saveAll(letterAnalyses);
-
-        return DailyReportResponseDto.of(dailyReport, letterAnalyses);
-    }
-
-    public DailyReportResponseDto getDailyReport(String userId, LocalDate targetDate) {
-        DailyReport dailyReport = dailyReportRepository.findByUserAndTargetDate(UUID.fromString(userId), targetDate)
-                .orElseThrow(
-                        () -> new DailyReportNotFoundException("Daily Report not found. targetDate: " + targetDate));
-
-        List<LetterAnalysis> letterAnalyses = letterAnalysisRepository.findByDailyReportId(dailyReport.getId());
-
-        return DailyReportResponseDto.of(dailyReport, letterAnalyses);
-    }
-
-    /**
-     * <ol> userId와 targetDate에 해당하는 편지를 최대 3건 조회합니다.
-     *     <li>만약 오늘이라면, 현재 시점 기준 가장 최근 3건</li>
-     *     <li>만약 오늘 이전이라면, 해당 날짜의 가장 마지막 3건</li>
-     * </ol>
-     *
-     * @param userId     일일분석 요청하는 사용자 아이디
-     * @param targetDate 일일분석 요청할 날짜
-     * @return 편지 3건 리스트
-     */
-    private List<Letter> findRecentLetters(UUID userId, LocalDate targetDate) {
-        LocalDateTime endTime = targetDate.isEqual(LocalDate.now())
-                ? LocalDateTime.now()
-                : targetDate.atTime(LocalTime.MAX);
-
-        List<Letter> letters = letterRepository.find3RecentLetters(userId, targetDate.atStartOfDay(), endTime);
-
-        if (letters.isEmpty()) {
-            throw new LetterNotFoundException("Letters for daily analysis not found.");
-        }
-        return letters;
-    }
-
-    private ClovaResponseDto requestClovaAnalysis(List<Letter> letters) {
-        String formattedMessages = letters.stream()
-                .map(Letter::getMessage)
-                .map(msg -> "\"" + msg + "\"")
-                .collect(Collectors.joining(", "));
-
-        return clovaService.sendDailyReportRequest(formattedMessages);
-    }
-
-    private DailyReport buildDailyReport(LocalDate targetDate, ClovaDailyAnalysisResult clovaDailyAnalysisResult) {
-        return DailyReport.builder()
-                .targetDate(targetDate)
-                .coreEmotion(CoreEmotion.valueOf(clovaDailyAnalysisResult.getDailyCoreEmotion()))
-                .description(clovaDailyAnalysisResult.getDescription())
-                .build();
-    }
-
-    private List<LetterAnalysis> buildLetterAnalyses(List<Letter> letters,
-                                                     ClovaDailyAnalysisResult clovaDailyAnalysisResult) {
-        return clovaDailyAnalysisResult.getLetterAnalyses().stream()
-                .map(analysis -> {
-                    int index = clovaDailyAnalysisResult.getLetterAnalyses().indexOf(analysis);
-                    Letter letter = letters.get(index); // 순서대로 letter 매핑
-
-                    return LetterAnalysis.builder()
-                            .letter(letter)
-                            .topic(analysis.getTopic())
-                            .coreEmotions(analysis.getCoreEmotions().stream()
-                                    .map(CoreEmotion::valueOf)
-                                    .collect(Collectors.toList()))
-                            .sensitiveEmotions(analysis.getSensitiveEmotions())
-                            .build();
-                })
-                .collect(Collectors.toList());
-    }
-
-    /*
-    - 분석 날짜, 작성빈도, 요일 ⛧ 감정 변화 추이, 위로 메세지, + 데일리의 각 편지 대표 감정
-    - 일요일 자정(00:00)을 넘으면 주간 분석 요청 버튼 활성화 및 주간 분석 요청 가능
+     * 분석 날짜, 작성빈도, 요일 ⛧ 감정 변화 추이, 위로 메세지, + 데일리의 각 편지 대표 감정
+     * 일요일 자정(00:00)을 넘으면 주간 분석 요청 버튼 활성화 및 주간 분석 요청 가능
      */
     public WeeklyReportResponseDto createWeeklyReport(WeeklyReportRequestDto weeklyReportRequestDto) {
         // 주간 분석 생성할 수 있는지 검증
@@ -258,16 +152,52 @@ public class ReportService {
         return WeeklyReportResponseDto.from(weeklyReport, manager);
     }
 
-    private List<LocalDate> createOneWeek(LocalDate startDate) {
-        return IntStream.range(0, 7)
-                .mapToObj(startDate::plusDays)
-                .toList();
-    }
-
     public WeeklyReportResponseDto getWeeklyReport(UUID userId, LocalDate startDate, LocalDate endDate) {
         WeeklyReport weeklyReport = weeklyReportRepository.findDailyReportBy(userId, startDate, endDate)
                 .orElseThrow(() -> new WeeklyReportNotFoundException("주간분석이 존재하지 않습니다"));
 
         return WeeklyReportResponseDto.from(weeklyReport);
+    }
+
+    private ClovaResponseDto requestClovaAnalysis(List<Letter> letters) {
+        String formattedMessages = letters.stream()
+                .map(Letter::getMessage)
+                .map(msg -> "\"" + msg + "\"")
+                .collect(Collectors.joining(", "));
+
+        return clovaService.sendDailyReportRequest(formattedMessages);
+    }
+
+    private DailyReport buildDailyReport(LocalDate targetDate, ClovaDailyAnalysisResult clovaDailyAnalysisResult) {
+        return DailyReport.builder()
+                .targetDate(targetDate)
+                .coreEmotion(CoreEmotion.valueOf(clovaDailyAnalysisResult.getDailyCoreEmotion()))
+                .description(clovaDailyAnalysisResult.getDescription())
+                .build();
+    }
+
+    private List<LetterAnalysis> buildLetterAnalyses(List<Letter> letters,
+                                                     ClovaDailyAnalysisResult clovaDailyAnalysisResult) {
+        return clovaDailyAnalysisResult.getLetterAnalyses().stream()
+                .map(analysis -> {
+                    int index = clovaDailyAnalysisResult.getLetterAnalyses().indexOf(analysis);
+                    Letter letter = letters.get(index); // 순서대로 letter 매핑
+
+                    return LetterAnalysis.builder()
+                            .letter(letter)
+                            .topic(analysis.getTopic())
+                            .coreEmotions(analysis.getCoreEmotions().stream()
+                                    .map(CoreEmotion::valueOf)
+                                    .collect(Collectors.toList()))
+                            .sensitiveEmotions(analysis.getSensitiveEmotions())
+                            .build();
+                })
+                .collect(Collectors.toList());
+    }
+
+    private List<LocalDate> createOneWeek(LocalDate startDate) {
+        return IntStream.range(0, 7)
+                .mapToObj(startDate::plusDays)
+                .toList();
     }
 }
